Minor refactor
- Move app and client to examples folder
- Standardize indentation to 4 spaces
