+++ src/example/rich-text-value.json
@@ -1,0 +1,106 @@
+{
+    "document": {
+        "nodes": [
+            {
+                "object": "block",
+                "type": "paragraph",
+                "nodes": [
+                    {
+                        "object": "text",
+                        "leaves": [
+                            {
+                                "text": "This is editable "
+                            },
+                            {
+                                "text": "rich",
+                                "marks": [
+                                    {
+                                        "type": "bold"
+                                    }
+                                ]
+                            },
+                            {
+                                "text": " text, "
+                            },
+                            {
+                                "text": "much",
+                                "marks": [
+                                    {
+                                        "type": "italic"
+                                    }
+                                ]
+                            },
+                            {
+                                "text": " better than a "
+                            },
+                            {
+                                "text": "<textarea>",
+                                "marks": [
+                                    {
+                                        "type": "code"
+                                    }
+                                ]
+                            },
+                            {
+                                "text": "!"
+                            }
+                        ]
+                    }
+                ]
+            },
+            {
+                "object": "block",
+                "type": "paragraph",
+                "nodes": [
+                    {
+                        "object": "text",
+                        "leaves": [
+                            {
+                                "text": "Since it's rich text, you can do things like turn a selection of text "
+                            },
+                            {
+                                "text": "bold",
+                                "marks": [
+                                    {
+                                        "type": "bold"
+                                    }
+                                ]
+                            },
+                            {
+                                "text": ", or add a semantically rendered block quote in the middle of the page, like this:"
+                            }
+                        ]
+                    }
+                ]
+            },
+            {
+                "object": "block",
+                "type": "block-quote",
+                "nodes": [
+                    {
+                        "object": "text",
+                        "leaves": [
+                            {
+                                "text": "A wise quote."
+                            }
+                        ]
+                    }
+                ]
+            },
+            {
+                "object": "block",
+                "type": "paragraph",
+                "nodes": [
+                    {
+                        "object": "text",
+                        "leaves": [
+                            {
+                                "text": "Try it out for yourself!"
+                            }
+                        ]
+                    }
+                ]
+            }
+        ]
+    }
+}
