+++ src/index.js
@@ -1,7 +1,7 @@
+import './index.css';
+import App from './example/App';
 import React from 'react';
 import ReactDOM from 'react-dom';
-import './index.css';
-import App from './App';
 import registerServiceWorker from './registerServiceWorker';
 
 ReactDOM.render(<App />, document.getElementById('root'));
